Replace pseudo slanted characters
* font/devpdf/DESC.in:
* font/devpdf/SS:
* font/devpdf/StandardSymSL.pfb:
* tmac/pdf.tmac:

Preparation for replacing the pseudo slanted lowercase greek
characters (used in equations) with real glyphs, i.e. provide a
Symbol-Slanted pfb font for gropdf. At the moment it requires the
following intervention to make it work with test-groff:-

Copy files SS and StandardSymSL.pfb to the build directory.

Add the following line to the download file:-

        Symbol-Slanted  ./StandardSymSL.pfb.

This needs to be added to the devpdf.am makefile so that it happens
during make, and ensure the two new files are copied to the install
directory and are cleaned and uninstalled appropriately. The macro
pdf:tmac which created the pseudo slanted glyphs previously has been
removed.

Without the intervention, it may appear that all is working when
test-groff is run (looking a pdf of groff_char(7) the lower greek
characters may be slanted, but if you look at the fonts used you will
see that the slanted symbol font has been substituted by some italic
font from the system), and also there will be problems when groff is
run from its install directory.

So don't cherry-pick this commit until the makefile has been suitably
amended.