Problem processing concatenated dit files (-Z)
Normally it is safe to pass concatenated dit files to gropdf,
typically this would be a separate source file which produces a
custom cover and a different source for the body (different macro
set?). Problem occurs if one of the dits uses the same fontno for
a different font. Normally troff allocates TR to #5, but if one
of the dits has been run with the flag "U-T" then U-TR is allocated
to #5.

* src/devices/gropdf/gropdf.pl: Check if "x font # name" has the
same number AND name as a previously registered font, otherwise
reload the font.