[tbl,tests]: Apply recent idiom consistently.
We weren't testing the shell variable `fail` for emptiness at the end of
the script.  This didn't matter becuse there's only one check in this
script anyway, and it was exiting with the status of the relevant
command, but if we add more checks it could cause misleading results.
(For example, if check 1 failed but check 2 succeeded, the script would
have falsely reported a pass.)