Distribute bootstrap and bootstrap.conf
With the recent xz-utils backdoor, there's been more focus on cases
where build systems rely on files produced by "make dist" and included
in release tarballs.  It's already fairly standard practice for
distributions to rebuild configure scripts using autoreconf, but less so
to rebuild the files that are produced by gnulib.

I looked into what it would take for Debian's groff package to do a full
rebootstrap from its packaged version of gnulib.  It seems relatively
straightforward, but it requires including bootstrap and bootstrap.conf
in tarballs so that we know what modules to use.

I've omitted README.git to ensure that we still warn people who don't
know what they're doing that running "./bootstrap" may not be the right
place to start.

* Makefile.am (EXTRA_DIST): Add "bootstrap" and "bootstrap.conf".