[man]: Fix Savannah #65729 (`de1 EE`).
* tmac/an.tmac (EE): Define macro with `de1` request, not `de`.

Fixes <https://savannah.gnu.org/bugs/?65729>.  Thanks to Bjarni Ingi
Gislason for the report.  Problem introduced by me in commit 15f8188656,
21 February 2022.

As a comment in the file notes:

Macros that are part of the external interface (TH, SH, P, etc.) or that
are called by traps of any kind must be defined with `de1` because they
might be called from a context where compatibility mode is enabled.  For
other macros, `de` suffices.