[troff]: Fix failing test.
* src/roff/troff/mtsm.cpp (mtsm::inherit): Fix failing test.

I freely admit I don't completely understand why this works.  But it
restores correct man page formatting.

Complicating the problem is that grohtml/MTSM's support for centering
and right-alignment (the `ce` and `rj` requests) never really worked
anyway.

Try this with any version of groff supporting grohtml:

$ cat ATTIC/ce-rj-html.groff
.ce 1
This is centered.
This is not.
.rj 1
This is right-aligned.
This is aligned to the left.
$ ~/groff-1.22.3/bin/groff -Tascii ./ATTIC/ce-rj-html.groff | cat -s
                        This is centered.
This is not.
                                           This is right-aligned.
This is aligned to the left.
$ ~/groff-1.22.3/bin/groff -Thtml ./ATTIC/ce-rj-html.groff | grep '<p'
<p align="center">This is centered. This is not.</p>
<p>This is right-aligned. This is aligned to the left.</p>

Why were these features (in grohtml/MTSM) put in if they didn't work?