Revert "[troff]: Slightly refactor escape seq handling."
This reverts commit 2fa56ec84577fcc37d6f38f650b3e42a2d072bd5.

I found a case where we can in fact enter this function with a newline
as the next token.  Unfortunately, it is hard to find a minimal
reproducer for it.  But here is the meat of the case.

.nr li*prefix-pad 1
.de XX
.ds li*c-mark \" empty
.if \\n[.$]>0          .as li*c-mark \\$1\"
.if \\n[li*prefix-pad] .as li*c-mark \ \"
.ie d li*enumerator    .as li*c-mark \\*[li*enumerator}\" XXX
.el                    .as li*c-mark \\*[li*mark]\"
..
.XX
.XX *

The line flagged "XXX" causes the reverted assertion to fail when the
foregoing code is applied to the `LI` macro in our mm package.