groff_mm(7): Fix terminology clash.
The "type" used to describe the first argument to the `AL` macro is not
the "type" used by the `LB` macro that `AL` then calls.

AL's first argument is a register format (see the `af` request), and
LB's "type" determines how that number is decorated (with brackets, a
trailing dot, and so forth).