[grohtml]: Fix Savannah #65960 (s/pnmcut/pamcut/).
Search for and use pamcut(1), not pnmcut(1).  Per upstream, the former
was introduced 23 years ago and the latter withdrawn 15 years ago.  And
we just got the memo!

* m4/groff.m4 (GROFF_CHECK_GROHTML_PROGRAMS):
* src/preproc/html/pre-html.cpp (imageList::createImage): Do it.

* src/roff/groff/tests/html_works_with_grn_and_eqn.sh:
* src/roff/groff/tests/smoke-test_html_device.sh: Update tests.

* NEWS:
* README.MinGW:
* src/devices/grohtml/grohtml.1.man (Dependencies): Document it.

Fixes <https://savannah.gnu.org/bugs/?65960>.  Thanks to Bjarni Ingi
Gislason for the report.

In principle, to support really old systems we could parameterize the
name of the p*mcut executable.  But I don't want to add that complexity
until and unless sufficient demand is evident.