src/roff/troff/node.{h,cpp}: Update annotations.
Annotate null pointers with `nullptr` comment to ease any future
transition to C++11, which defines it as a keyword.

Also add explanatory comments to formal arguments taking default values,
as a compromise with the Stroustrup-style C++ used in most of groff.

Some of the formal argument names are indeed not very "explanatory".
This is an area for improvement.