src/roff/troff/reg.cpp: Fix code style nits.
* src/roff/troff/reg.cpp (number_value_to_ascii): Add parentheses to
  complex expressions.

  (number_value_to_ascii): Assign character, not integer, literal to
  element of array of `char` type.

  (inline_define_register [0], define_register, set_register)
  (look_up_register, alter_format): Use `static_cast` instead of C-style
  cast on dictionary values.

  (inline_define_register [0], define_register, set_register)
  (look_up_register, alter_format): Reorder equality comparisons to
  avoid inadvertent lvalue assignment.

Use `sizeof` operator idiomatically; it is an operator, not a function,
and should be followed by parentheses only when making it operate on a
type cast to get the size of a non-lvalue.

Annotate null pointers with `nullptr` comment to ease any future
transition to C++11, which defines it as a keyword.