src/roff/troff/node.cpp: Solve commented mystery.
The answer to the XXX question posed is that a garbage character could
appear anywhere inside a lengthy escape sequence parameter.

$ printf '\\[`\001e]\n.pl 1v\n' | nroff
troff:<standard input>:1: error: a leader character is not allowed in an escape sequence parameter
e]

Above, the leader lies between two valid characters.  Take it out and
all is well.

$ printf '\\[`e]\n.pl 1v\n' | nroff
è