[troff]: Trivially refactor, reform terminology.
[troff]: Call `\X` escape sequence and `device` request device
_extension_ instructions.  There is already, in groff, a device command
'x u' that configures underlining of spaces (and which is implemented
only by grotty(1)).  A dusty corner, to be sure, but we could support
others in the future, and further, the new term indicates the fact that
`\X` and `device` produce arguments only for 'x X' "sub"-commands--this
has always been the case and for `\X` it was true even in Kernighan
troff.

* src/roff/troff/input.cpp (do_device_control): Rename this...
  (do_device_extension): ...to this.

  (do_device_extension, device_request): Migrate diagnostic
  messages.

* src/roff/troff/troff.1.man (Options): Migrate terminology.