[me]: Add FSF copyright notice.
The beginning of the year range is a guess on GBR's part.  The earliest
mention of groff me(7) in surviving documentation is in the "NEWS"
file's entry for groff 1.01, implying a GNU-specific change to something
that was already present.  groff 1.02 was released June 1991.  Wikipedia
documents (with no supporting citation) that groff 0.3.1 was released
one year earlier, in June 1990.  Other groff copyright notices for
portions of the source that go back to "the beginning" start with a year
in 1989, and since me(7), which was Free Software, would have been
useful to James Clark when verifying the operation of his formatter, I
surmise that he included it from very early days.  I welcome correction.

Also be aware of this paragraph from our "LICENSES" file:

"Some files are part of groff but bear licenses in addition to the GPL,
or have been placed into the public domain, because they originated
elsewhere; often, the groff project has modified them, sometimes
extensively."

As far as GBR is concerned, his changes to groff me(7) are disjunctively
("dual-") licensed under the GNU GPLv3 and under the BSD license under
which Clark obtained it, so that other me(7)s may feel free to copy our
enhancements to retain compatibility.  GBR infers that this has long
been the intention of GNU contributors' changes to me(7): groff me(7) is
careful to gate its use of groff extensions behind `.if \n(.g` tests.