[mm]: Handle `LB`'s fifth argument more flexibly.
* contrib/mm/m.tmac (LB): Accept a dummy character escape sequence as a
  synonym for an empty fifth argument.

* contrib/mm/groff_mm.7.man (Macros) <LB>: Document this.

This is to save the user from the tedium of figuring out how to escape
or quote an empty pair of double quotes as a macro argument in case they
indirectly assign to the (forthcoming) string that parameterizes the
reference list format.  The need for this is a knock-on effect of Bell's
antipathy for strong type systems.  Possibly the attitude is reflected
in a contemporaneous satire on advertising.

"Hey, hey, hey!  Calm down, you two--New Shimmer is both a floor wax
_and_ a dessert topping!" -- _Saturday Night Live_, 1976-01-10

(Holmdel was getting up to mm mischief at about this time.)