[tbl]: Refactor for less inefficiency.
* src/preproc/tbl/table.cpp (table::add_entry): Refactor.  Since we use
  the C string `extract()`ion of the groff `string` table entry
  repeatedly, move logic that stores its pointer to an automatic
  variable much earlier, and reference that in diagnostic messages
  instead of repeatedly calling a member function.

A thing to know about groff's `string::extract()` is that it allocates
memory.[1]  One might wonder, as I did, when that memory is ever freed.
The answer is: when tbl(1) exits.  That's because tbl needs these string
pointers after `table::add_entry()` returns, or there won't be much in
your table.  Ah, to program like it's 1989 again.

A more fastidious approach would copy the pointers to each of these
allocations into an STL `vector` (possibly a private member of class
`table`), and then, when a table region is finished (`.T&` or `.TE` is
encountered), walk the vector and free the storage.

[1] https://git.savannah.gnu.org/cgit/groff.git/tree/src/libs/libgroff/string.cpp?h=1.23.0#n285