[mm]: Fix Savannah #66287 (2/2) (coding tag).
* contrib/mm/groff_mmse.7.man: Add old-style Emacs file-local variable
  to supply preconv(1) a "coding tag".  Doing so throws that program a
  bone if the "uchardet" library is not available and the program is
  asked to guess the file's input encoding.

Fixes <https://savannah.gnu.org/bugs/?66287> (2/2).  Thanks to Dave
Kemper for the report.

Also update the command examples for readers interested in rendering the
file.  The examples specify the input encoding explicitly; the coding
tag is for the benefit of people who use the groff(1) option "-k"
instead of "-K".

This commit causes a preconv "smoke test" failure.