tmac/{de,fr}.tmac: Respell string translations.
* tmac/de.tmac:
* tmac/fr.tmac: Spell string translations using groff special character
  escape sequences instead of Latin-1 or Latin-9 code points; this way
  they work with a document that uses them no matter what its own
  encoding.

Cf. commit 944ca3d436 (23 October), which did the same for "sv.tmac".

I didn't do "ru.tmac"; that one's more of a pickle.  The goal identified
above could be achieved by sifting the string translations through
preconv(1) and committing that, but that would come at the cost of
rendering them unintelligible to humans (and therefore prone to error).
We may have to kick that can down the road beyond the "GNU troff accepts
UTF-8 input" milestone.