[doc,man]: Clarify behavior of pi and sy requests.
Refer to each request's argument as "contents" to indicate for
consistency with the way string-populating requests are handled, and to
aid the reader to recall that the argument will be read to the end of
the input line in copy mode.

There's one difference: `pi` and `sy` don't handle a leading double
quote the way `ds` and `as` do.  But I'm about to change that.

* doc/groff.texi.in: Bump document date.