[indxbib]: Improve header file hygiene.
* src/utils/indxbib/indxbib.cpp: Include local "nonposix.h" header file,
  to try to ensure visibility of `getcwd()` symbol on non-POSIX systems.

The only such declaration present there is conditioned on `__EMX__`
though, which for those born after 1990, was a (sort of?) POSIX layer
for IBM OS/2.  This shouldn't make the situation _worse_.  Why haven't
the numerous OS/2 users indexing their bibliographic reference files to
speed up refer(1) preprocessing with groff in 2024 complained to us?