pic(1): Be more frugal with vertical space.
This unstrands a line in U.S. letter paper format.

I'm not sure that an example that is inlined into a sentence should be
displayed.  By all means, if it's more than one line, indent it and
change the font family, but having vertical space around it seems a bit
hinky to me.