[nroff]: Improve shell behavior check.
...and fix goof in test script.

* src/roff/nroff/nroff.sh: Replace proxy `unset`-based test with an
  actual measurement of parameter expansion.

* src/roff/nroff/tests/verbose_option_works.sh: Fix spurious failure of
  test for grep(1) supporting `-q` and `-x` options.  Continues commit
  e05cf6d3b3, 4 December.

Every time I have to deal with Solaris, it's like gazing into the Abyss.

A test failure of "src/roff/nroff/tests/verbose_option_works.sh" is
expected on Solaris 10, because 2 of the checks in that script verify
that option clustering works, and with Solaris 10 /bin/sh, it doesn't.
The test failure is therefore not spurious.  One can edit the installed
nroff script in place to use "/usr/xpg4/bin" in its shebang line.