Revamp build-time check of test(1) command.
...for POSIX conformance.

"test -ef", long depended upon by the contributed program gdiffmk(1) for
certain functionality, is now standard as of POSIX Issue 8 (2024).  The
Austin Group's threshold criteria for inclusion in the standard mean
that such a feature is widely available, and one need not
pessimistically assume that the Bash shell must be used to obtain the
feature.  Refactor and revise our detection and fallback techniques
accordingly.

* m4/groff.m4 (GROFF_HAVE_TEST_EF_OPTION): Rename this...
  (GROFF_PROG_TEST_SUPPORTS_EF_OPTION): ...to this.  Simplify,
  making it a simple test of the feature, using "sh -c" since we
  don't care whether the shell itself or a 'test' executable in
  the $PATH provides it.

  (GROFF_BASH): Rename this...
  (GROFF_PROG_SH_IS_POSIX_8_CONFORMING): ...to this.  It's a
  mouthful, but reflects our concerns.  We don't care about Bash
  per se (groff does not employ Bashisms); we care about whether
  the shell conforms to POSIX Issue 8.  Rename the `AC_SUBST`ed
  macro `BASH_PROG` to `POSIX_SHELL_PROG`.  If we don't find a
  conforming shell, _then_ fall back to Bash for
  `POSIX_SHELL_PROG`, if it is available; if not, configure-time
  tests are fruitless and programs must perform runtime checks for
  POSIX non-conformance (since we're not shipping our own shell or
  test(1) to close the gap).

* configure.ac: Update expansion sites of m4 macros to use their new
  names.

Tests of `gdiffmk` may fail at this commit.  (They don't on GBR's
system, but the generated "gdiffmk" program text is not quite correct.)