[mdoc]: Support C23 in the .St macro.
* tmac/mdoc/doc-syms: Support .St -isoC-2023.
* tmac/groff_mdoc.7.man: Document .St -isoC-2023.

The mismatch of the year numbers is not a typo.
The official name is "ISO/IEC 9899:2024", so the "4" is correct there.

But as a colloquial name, "C23" is more widespread than "C24",
probably because __STDC_VERSION__ == 202311L: The final version
of the standard was ready in 2023, only formally publishing it
took about 11 months after it was ready.

The main reason for choosing "-isoC-2023" as the .St argument is to
match what NetBSD, OpenBSD and mandoc already committed to their trees.
Also, it aligns well with the existing "-isoC-2011".

Not adding -isoC-2017 for C17 is intentional because that edition only
fixes bugs and adds no features, so it's hard to see why anybody would
want to reference it in a manual page.  Keeping the code and the
documentation concise by avoiding needless arguments feels preferable.

Addition suggested by Thomas Klausner <wiz at NetBSD dot org>.

"No objection on my part."  gbranden@