groff_man*(7): Reform typographical convention.
Include the leading dot before a macro name in only two situations: (1)
examples, and (2) the presentation defining the macro's meaning.  (The
latter happens in two places; the one-line summary serving as a sort of
directory or quick reference at the top of the document, and the
conventioned tagged paragraphs synopsizing usage.)

My objectives are twofold: (A) Help the reader to understand that the
dot is not part of the macro's name; (B) help the reader searching the
document for a definitions of a macro or examples of its use to find
them more easily, without a bunch of false positives getting in the way.