[troff]: Fix code style nit.
* src/roff/troff/node.cpp (hline_node::~hline_node)
  (vline_node::~vline_node): Fix code style nit, simplifying.

It is not necessary to make conditional the `delete[]` of a null
pointer.  "If the _delete-expression_ calls the implementation
deallocation function (3.7.3.2), and if the operand of the delete
expression is not the null pointer constant, the deallocation function
will deallocate the storage referenced by the pointer thus rendering the
pointer invalid" (ISO/IEC 14882-1998, §5.3.5, paragraph 4).  Or as
Stroustrup puts it, "Applying _delete_ to zero has no effect." (_The C++
Programming Language, Special Edition_, p. 128).