[troff]: Implement recursive node dumping (5b/9).
* src/roff/troff/node.h (class diverted_space_node): Specialize
  (override) `dump_properties()` for this class.

* src/roff/troff/node.cpp (diverted_space_node::dump_properties): New
  member function reports value of `vunits` property.

At present I don't know of a way to get a diverted space node to report
itself in debugging output; the `pline` request works only on nodes that
are in a pending output line, and a diverted space node, being a form of
vertical space (and not "local" vertical motion), exists only between
output lines.  Nevertheless this plants a tree today whose shade we
might enjoy later.