[troff]: Implement recursive node dumping (5c/9).
* src/roff/troff/node.h (class diverted_copy_file_node): Specialize
  (override) `dump_properties()` for this class.

* src/roff/troff/node.cpp (diverted_copy_file_node::dump_properties):
  New member function reports value of `filename` property.

At present I don't know of a way to get a diverted copy file node to
report itself in debugging output; the `pline` request works only on
nodes that are in a pending output line, and nodes of this type are
added only to a `macro_diversion` object, not an output line node list.
Nevertheless this plants a tree today whose shade we might enjoy later.

Here's an example of what its output would look like:

{"type": "diverted_copy_file_node", "diversion level": 0, "is_special_node": false, "filename": "ATTIC/trf-fodder"}