[troff]: Implement recursive node dumping (5e/9).
* src/roff/troff/node.h (class vertical_size_node): Specialize
  (override) `dump_properties()` for this class.

* src/roff/troff/node.cpp (vertical_size_node::dump_properties): New
  member function reports value of `vunits` property.

At present I don't know of a way to get a vertical size node to report
itself in debugging output; the `pline` request works only on nodes that
are in a pending output line, and nodes of this type are added only to a
`macro_diversion` object, not an output line node list.  Nevertheless
this plants a tree today whose shade we might enjoy later.