[troff]: Implement recursive node dumping (5q/9).
* src/roff/troff/node.h (class tag_node): Specialize (override)
  `dump_properties()` for this class.

* src/roff/troff/node.cpp (tag_node::dump_properties): New member
  function reports values of `string` and `delayed` properties.

"Tag nodes", like "suppress nodes", are some of groff's blacker magic,
and are entangled with grohtml.  Given the following simple input
document:

.NH 1
Introduction
.pline
.LP
Hello tag node world.

...processed with `groff -z -ms -Thtml`, this commit changes `pline`
request output as follows.

-{"type": "tag_node", "diversion level": 0, "is_special_node": true, "state": "<state [br] [sp 1]>"},
+{"type": "tag_node", "diversion level": 0, "is_special_node": true, "state": "<state [br] [sp 1]>", "string": "x X devtag:.NH 1\n", "delayed": false},

It's worth noting that the "string" gets populated with a trailing
newline (this is impossible for *roff strings, but this `string` type
is implemented as a C++ class).  This may or may not be a bug, depending
on how tag appending works.  If it's not a bug, we may want to rename
this property to something other than "string" to reduce potential user
confusion.